# logcollector/openapi.json
{"openapi":"3.1.0","info":{"title":"LogParser (Refactored)","version":"2.0.0"},"paths":{"/api/parse":{"post":{"summary":"Parse Yaml","description":"Endpoint test pour parser YAML manuellement","operationId":"parse_yaml_api_parse_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParseRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/parse/db":{"post":{"summary":"Parse From Db","description":"Récupère un log depuis la collection 'log' de MongoDB et le parse.","operationId":"parse_from_db_api_parse_db_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DBParseRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"DBParseRequest":{"properties":{"run_id":{"type":"string","title":"Run Id"},"provider":{"type":"string","title":"Provider","default":"GITHUB"}},"type":"object","required":["run_id"],"title":"DBParseRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ParseRequest":{"properties":{"yaml":{"type":"string","title":"Yaml"}},"type":"object","required":["yaml"],"title":"ParseRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}